Add xcvm.rs
