+++ runtime/unique/src/opaque.rs
@@ -6,8 +6,6 @@
 use sp_runtime::impl_opaque_keys;
 use sp_std::prelude::Vec;
 
-pub use unique_runtime_common::types::{Block, UncheckedExtrinsic};
-
 use super::constructor::Aura;
 
 impl_opaque_keys! {
