+++ runtime/unique/src/xcvm.rs
@@ -1,0 +1,237 @@
+use cumulus_pallet_xcm;
+use frame_support::{match_types, parameter_types, weights::Weight};
+use frame_support::pallet_prelude::Get;
+use frame_support::traits::{Currency as CurrencyT, Everything, OnUnbalanced as OnUnbalancedT};
+use frame_support::weights::{WeightToFee, WeightToFeeCoefficient, WeightToFeeCoefficients, WeightToFeePolynomial};
+use pallet_xcm::XcmPassthrough;
+use polkadot_parachain::primitives::Sibling;
+use sp_arithmetic::traits::{SaturatedConversion, Saturating};
+use sp_runtime::traits::{CheckedConversion, Zero};
+use sp_std::marker::PhantomData;
+use xcm::latest::{Error as XcmError, MultiAsset};
+use xcm::prelude::{Concrete, Fungible as XcmFungible};
+use xcm::v1::{BodyId, Junction::{Parachain, Plurality}, Junctions::*, MultiLocation, NetworkId};
+use xcm_builder::{AccountId32Aliases, AllowTopLevelPaidExecutionFrom, AllowUnpaidExecutionFrom, CurrencyAdapter, FixedWeightBounds, LocationInverter, NativeAsset, ParentAsSuperuser, ParentIsPreset, RelayChainAsNative, SiblingParachainAsNative, SiblingParachainConvertsVia, SignedAccountId32AsNative, SignedToAccountId32, SovereignSignedViaLocation, TakeWeightCredit};
+use xcm_executor::{Assets, Config};
+use xcm_executor::traits::{MatchesFungible, WeightTrader};
+
+use unique_runtime_common::{constants::{MAXIMUM_BLOCK_WEIGHT, UNIQUE}, types::{AccountId, Balance}};
+
+use crate::LinearFee;
+
+use super::constructor::{Balances, Call, Origin, ParachainInfo, ParachainSystem, PolkadotXcm, XcmpQueue};
+
+parameter_types! {
+	pub const ReservedDmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT / 4;
+	pub const ReservedXcmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT / 4;
+
+	pub const RelayLocation: MultiLocation = MultiLocation::parent();
+	pub const RelayNetwork: NetworkId = NetworkId::Polkadot;
+	pub RelayOrigin: Origin = cumulus_pallet_xcm::Origin::Relay.into();
+	pub Ancestry: MultiLocation = Parachain(ParachainInfo::parachain_id().into()).into();
+}
+
+/// Type for specifying how a `MultiLocation` can be converted into an `AccountId`. This is used
+/// when determining ownership of accounts for asset transacting and when attempting to use XCM
+/// `Transact` in order to determine the dispatch Origin.
+pub type LocationToAccountId = (
+    // The parent (Relay-chain) origin converts to the default `AccountId`.
+    ParentIsPreset<AccountId>,
+    // Sibling parachain origins convert to AccountId via the `ParaId::into`.
+    SiblingParachainConvertsVia<Sibling, AccountId>,
+    // Straight up local `AccountId32` origins just alias directly to `AccountId`.
+    AccountId32Aliases<RelayNetwork, AccountId>,
+);
+
+pub struct OnlySelfCurrency;
+
+impl<B: TryFrom<u128>> MatchesFungible<B> for OnlySelfCurrency {
+    fn matches_fungible(a: &MultiAsset) -> Option<B> {
+        match (&a.id, &a.fun) {
+            (Concrete(_), XcmFungible(ref amount)) => CheckedConversion::checked_from(*amount),
+            _ => None,
+        }
+    }
+}
+
+/// Means for transacting assets on this chain.
+pub type LocalAssetTransactor = CurrencyAdapter<
+    // Use this currency:
+    Balances,
+    // Use this currency when it is a fungible asset matching the given location or name:
+    OnlySelfCurrency,
+    // Do a simple punn to convert an AccountId32 MultiLocation into a native chain account ID:
+    LocationToAccountId,
+    // Our chain's account ID type (we can't get away without mentioning it explicitly):
+    AccountId,
+    // We don't track any teleports.
+    (),
+>;
+
+/// This is the type we use to convert an (incoming) XCM origin into a local `Origin` instance,
+/// ready for dispatching a transaction with Xcm's `Transact`. There is an `OriginKind` which can
+/// biases the kind of local `Origin` it will become.
+pub type XcmOriginToTransactDispatchOrigin = (
+    // Sovereign account converter; this attempts to derive an `AccountId` from the origin location
+    // using `LocationToAccountId` and then turn that into the usual `Signed` origin. Useful for
+    // foreign chains who want to have a local sovereign account on this chain which they control.
+    SovereignSignedViaLocation<LocationToAccountId, Origin>,
+    // Native converter for Relay-chain (Parent) location; will converts to a `Relay` origin when
+    // recognised.
+    RelayChainAsNative<RelayOrigin, Origin>,
+    // Native converter for sibling Parachains; will convert to a `SiblingPara` origin when
+    // recognised.
+    SiblingParachainAsNative<cumulus_pallet_xcm::Origin, Origin>,
+    // Superuser converter for the Relay-chain (Parent) location. This will allow it to issue a
+    // transaction from the Root origin.
+    ParentAsSuperuser<Origin>,
+    // Native signed account converter; this just converts an `AccountId32` origin into a normal
+    // `Origin::Signed` origin of the same 32-byte value.
+    SignedAccountId32AsNative<RelayNetwork, Origin>,
+    // Xcm origins can be represented natively under the Xcm pallet's Xcm origin.
+    XcmPassthrough<Origin>,
+);
+
+parameter_types! {
+	// One XCM operation is 1_000_000 weight - almost certainly a conservative estimate.
+	pub UnitWeightCost: Weight = 1_000_000;
+	// 1200 UNIQUEs buy 1 second of weight.
+	pub const WeightPrice: (MultiLocation, u128) = (MultiLocation::parent(), 1_200 * UNIQUE);
+	pub const MaxInstructions: u32 = 100;
+	pub const MaxAuthorities: u32 = 100_000;
+}
+
+match_types! {
+	pub type ParentOrParentsUnitPlurality: impl Contains<MultiLocation> = {
+		MultiLocation { parents: 1, interior: Here } |
+		MultiLocation { parents: 1, interior: X1(Plurality { id: BodyId::Unit, .. }) }
+	};
+}
+
+pub type Barrier = (
+    TakeWeightCredit,
+    AllowTopLevelPaidExecutionFrom<Everything>,
+    AllowUnpaidExecutionFrom<ParentOrParentsUnitPlurality>,
+    // ^^^ Parent & its unit plurality gets free execution
+);
+
+pub struct UsingOnlySelfCurrencyComponents<
+    WeightToFee: WeightToFeePolynomial<Balance=Currency::Balance>,
+    AssetId: Get<MultiLocation>,
+    AccountId,
+    Currency: CurrencyT<AccountId>,
+    OnUnbalanced: OnUnbalancedT<Currency::NegativeImbalance>,
+>(
+    Weight,
+    Currency::Balance,
+    PhantomData<(WeightToFee, AssetId, AccountId, Currency, OnUnbalanced)>,
+);
+
+impl<
+    WeightToFee: WeightToFeePolynomial<Balance=Currency::Balance>,
+    AssetId: Get<MultiLocation>,
+    AccountId,
+    Currency: CurrencyT<AccountId>,
+    OnUnbalanced: OnUnbalancedT<Currency::NegativeImbalance>,
+> WeightTrader
+for UsingOnlySelfCurrencyComponents<WeightToFee, AssetId, AccountId, Currency, OnUnbalanced>
+{
+    fn new() -> Self {
+        Self(0, Zero::zero(), PhantomData)
+    }
+
+    fn buy_weight(&mut self, weight: Weight, payment: Assets) -> Result<Assets, XcmError> {
+        let amount = WeightToFee::weight_to_fee(&weight);
+        let u128_amount: u128 = amount.try_into().map_err(|_| XcmError::Overflow)?;
+
+        // location to this parachain through relay chain
+        let option1: xcm::v1::AssetId = Concrete(MultiLocation {
+            parents: 1,
+            interior: X1(Parachain(ParachainInfo::parachain_id().into())),
+        });
+        // direct location
+        let option2: xcm::v1::AssetId = Concrete(MultiLocation {
+            parents: 0,
+            interior: Here,
+        });
+
+        let required = if payment.fungible.contains_key(&option1) {
+            (option1, u128_amount).into()
+        } else if payment.fungible.contains_key(&option2) {
+            (option2, u128_amount).into()
+        } else {
+            (Concrete(MultiLocation::default()), u128_amount).into()
+        };
+
+        let unused = payment
+            .checked_sub(required)
+            .map_err(|_| XcmError::TooExpensive)?;
+        self.0 = self.0.saturating_add(weight);
+        self.1 = self.1.saturating_add(amount);
+        Ok(unused)
+    }
+
+    fn refund_weight(&mut self, weight: Weight) -> Option<MultiAsset> {
+        let weight = weight.min(self.0);
+        let amount = WeightToFee::weight_to_fee(&weight);
+        self.0 -= weight;
+        self.1 = self.1.saturating_sub(amount);
+        let amount: u128 = amount.saturated_into();
+        if amount > 0 {
+            Some((AssetId::get(), amount).into())
+        } else {
+            None
+        }
+    }
+}
+
+impl<
+    WeightToFee: WeightToFeePolynomial<Balance=Currency::Balance>,
+    AssetId: Get<MultiLocation>,
+    AccountId,
+    Currency: CurrencyT<AccountId>,
+    OnUnbalanced: OnUnbalancedT<Currency::NegativeImbalance>,
+>
+Drop for UsingOnlySelfCurrencyComponents<WeightToFee, AssetId, AccountId, Currency, OnUnbalanced>
+{
+    fn drop(&mut self) {
+        OnUnbalanced::on_unbalanced(Currency::issue(self.1));
+    }
+}
+
+pub struct XcmConfig;
+
+impl Config for XcmConfig {
+    type Call = Call;
+    type XcmSender = XcmRouter;
+    // How to withdraw and deposit an asset.
+    type AssetTransactor = LocalAssetTransactor;
+    type OriginConverter = XcmOriginToTransactDispatchOrigin;
+    type IsReserve = NativeAsset;
+    type IsTeleporter = ();
+    // Teleportation is disabled
+    type LocationInverter = LocationInverter<Ancestry>;
+    type Barrier = Barrier;
+    type Weigher = FixedWeightBounds<UnitWeightCost, Call, MaxInstructions>;
+    type Trader = UsingOnlySelfCurrencyComponents<
+        LinearFee<Balance>, RelayLocation, AccountId, Balances, ()
+    >;
+    type ResponseHandler = ();
+    // Don't handle responses for now.
+    type AssetTrap = PolkadotXcm;
+
+    type AssetClaims = PolkadotXcm;
+    type SubscriptionService = PolkadotXcm;
+}
+
+/// No local origins on this chain are allowed to dispatch XCM sends/executions.
+pub type LocalOriginToLocation = (SignedToAccountId32<Origin, AccountId, RelayNetwork>, );
+
+/// The means for routing XCM messages which are not for local execution into the right message
+/// queues.
+pub type XcmRouter = (
+    // Two routers - use UMP to communicate with the relay chain:
+    cumulus_primitives_utility::ParentAsUmp<ParachainSystem, ()>,
+    // ..and XCMP to communicate with the sibling chains.
+    XcmpQueue,
+);
